+++ src/App.css
@@ -1,56 +1,3 @@
-button {
-  padding: 1.5rem 2.5rem;
-  font-size: 1.3rem;
-  font-weight: bold;
-  background-color: #fff;
-  border: 1px solid rgb(241 238 238);
-}
-
-.calc-body {
-  display: flex;
-  flex-flow: column;
-  list-style-type: none;
-  border: 2px solid rgb(241 238 238);
-  width: 40%;
+body {
   padding: 0;
-  margin: 150px auto 0;
-}
-
-.calc-screen {
-  text-align: end;
-  padding: 30px;
-  background-color: rgb(241 238 238);
-  font-weight: bold;
-  font-size: 1.5rem;
-}
-
-.calc-keyboard {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-  width: 100%;
-}
-
-.calc-digits {
-  display: grid;
-  flex: 1;
-  grid-template-columns: 1fr 1fr 1fr;
-}
-
-.calc-operators {
-  display: flex;
-  flex-flow: column;
-}
-
-.calc-operators button {
-  flex: 1;
-  margin: 1px;
-  background-color: orange;
-  border: 1px hidden;
-}
-
-.calc-digits button:nth-last-child(2) {
-  grid-column: span 2;
 }
